Separate env config from ley config
